Extract TX compression logic into separate file
